updating to match tutorial
