+++ tasks/jetpack-compose/.idea/misc.xml
@@ -3,6 +3,11 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
+        <entry key="../../../../../../layout/compose-model-1629961708702.xml" value="0.24444444444444444" />
+        <entry key="../../../../../../layout/compose-model-1630004115024.xml" value="0.15925925925925927" />
+        <entry key="../../../../../../layout/compose-model-1630005870269.xml" value="0.5" />
+        <entry key="../../../../../../layout/compose-model-1630009944964.xml" value="0.32586837294332727" />
+        <entry key="../../../../../../layout/compose-model-1630018155450.xml" value="0.45" />
         <entry key="../../../../../layout/compose-model-1629924662227.xml" value="0.47846283783783783" />
         <entry key="../../../../../layout/compose-model-1629929392169.xml" value="0.5" />
         <entry key="../../../../../layout/compose-model-1629937545393.xml" value="0.375" />
