+++ tasks/jetpack-compose/app/src/main/AndroidManifest.xml
@@ -14,12 +14,12 @@
     <uses-permission android:name="android.permission.CHANGE_WIFI_MULTICAST_STATE" />
 
     <application
+        android:name=".TasksApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:name=".TasksApplication"
         android:theme="@style/Theme.TasksJetpackCompose">
         <activity
             android:name=".MainActivity"
