+++ tasks/jetpack-compose/app/src/main/java/live/ditto/compose/tasks/MainActivity.kt
@@ -40,24 +40,5 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-@Composable
-fun Root() {
-    val navController = rememberNavController()
 
-    TasksJetpackComposeTheme {
-        // A surface container using the 'background' color from the theme
-        Surface(color = MaterialTheme.colors.background) {
-            NavHost(navController = navController, startDestination = "tasks") {
-                composable("tasks") { TasksListScreen(navController = navController) }
-                composable("tasks/edit") {
-                    EditScreen(navController = navController, taskId = null)
-                }
-                composable("tasks/edit/{taskId}") { backStackEntry ->
-                    val taskId: String? = backStackEntry.arguments?.getString("taskId")
-                    EditScreen(navController = navController, taskId = taskId)
-                }
-            }
-        }
-    }
-}
